requirements.txt
streamlitgit push -u origin main
pandas
numpy
scikit-learn
joblib
